llmspell_events::bus

Struct EventBus

pub struct EventBus { /* private fields */ }

Event bus for publishing and subscribing to events

Implementations

impl EventBus

pub fn new() -> Self

Create a new event bus with default configuration

pub fn with_config(config: FlowControllerConfig) -> Self

Create an event bus with custom flow control configuration

pub fn with_persistence<B: StorageBackend + 'static>( flow_config: FlowControllerConfig, storage_adapter: EventStorageAdapter<B>, persistence_config: PersistenceConfig, ) -> Self

Create an event bus with persistence using any storage backend

pub async fn publish(&self, event: UniversalEvent) -> Result<(), PublishError>

Publish an event to the bus

pub async fn subscribe( &self, pattern: &str, ) -> Result<UnboundedReceiver<UniversalEvent>, SubscribeError>

Subscribe to events matching a pattern

pub async fn subscribe_with_handler<H>( &self, pattern: &str, handler: H, ) -> Result<Uuid, SubscribeError>
where H: AsyncEventHandler + Send + 'static,

Subscribe with a custom event handler

pub fn subscribe_all(&self) -> Receiver<UniversalEvent>

Get a broadcast receiver for all events

pub fn buffer_size(&self) -> usize

Get current buffer size

pub fn get_stats(&self) -> FlowStats

Get flow statistics

pub fn subscription_count(&self) -> usize

Get number of active subscriptions

pub async fn get_persisted_events( &self, pattern: &str, ) -> Result<Vec<UniversalEvent>, Error>

Get persisted events by pattern (if persistence is enabled)

pub async fn get_events_by_correlation_id( &self, correlation_id: Uuid, ) -> Result<Vec<UniversalEvent>, Error>

Get persisted events by correlation ID (if persistence is enabled)

pub async fn get_storage_stats(&self) -> Result<Option<StorageStats>, Error>

Get storage statistics (if persistence is enabled)